update user api created

+++ src/main/java/com/iamsajan/examservice/service/impl/UserServiceImpl.java
@@ -2,6 +2,7 @@ package com.iamsajan.examservice.service.impl;
 
 import com.iamsajan.examservice.model.User;
 import com.iamsajan.examservice.model.UserRole;
+import com.iamsajan.examservice.model.exam.Quiz;
 import com.iamsajan.examservice.repository.RoleRepository;
 import com.iamsajan.examservice.repository.UserRepository;
 import com.iamsajan.examservice.service.UserService;
@@ -9,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 @Service
@@ -43,15 +45,37 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    public User getUserById(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser != null) return optionalUser.get();
+        return null;
+    }
+
+    @Override
+    public User updateUser(Long id, User user) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            User updatedUser = optionalUser.get();
+            updatedUser.setUsername(user.getUsername());
+            updatedUser.setFirstName(user.getFirstName());
+            updatedUser.setLastName(user.getLastName());
+            updatedUser.setGender(user.getGender());
+            updatedUser.setEmail(user.getEmail());
+            updatedUser.setPhone(user.getPhone());
+
+            return userRepository.save(updatedUser);
+        }
+        return null;
+    }
+
+    @Override
     public User getUserByUsername(String username) {
         return userRepository.findByUsername(username);
     }
 
     @Override
     public void deleteUserById(Long id) throws Exception {
-        if (userRepository.findById(id).isPresent())
-            userRepository.deleteById(id);
-        else
-            throw new Exception("User with id " + id + " doesn't exits !!!");
+        if (userRepository.findById(id).isPresent()) userRepository.deleteById(id);
+        else throw new Exception("User with id " + id + " doesn't exits !!!");
     }
 }
